Add magit-gitflow and fix the `f` issue for feature
The solution to fix the `f` binding issue is to not
alter emacs-state-map. By keeping emacs map untouched
we have a fallback for modes like magit-gitflow which
require the `f` key.

Indeed with the current implementation of the `fd`
sequence it is not possible to guess what the `f` key
must do for a given mode. The only solution for now is
to keep one not modified state map (emacs is a good candidate
for this).