Integrate ahs with evil search functionality
When auto-highlight-symbol is used, it does not currently place the
searched term into the search ring, so that after ahs mode is left,
hitting "n" or "N" does not return to the search term that was selected
by hitting "*" or "#".  Some extra steps must be taken to perform the
expected search behavior of "n" and "N" after "*" or "#".

This also integrates the searched term into the ex mode substitute
command, so that "%s//replace/" works as expected.