Update docs with more information on powerline colors
The powerline color issue is due to the fact that Emacs 24.4 now
defaults to using sRGB colorspace on OS X 10.7 and higher, not some
"brokeness" with brew's `emacs` build as the previous verbiage seemed to
imply; the powerline codebase is still assuming `(setq
ns-use-srgb-colorspace nil)` is the default.

This PR adds more information about this so that users may
decide whether they want to just set it back to nil, switch to
emacs-mac (which has it set to nil anyway), or disable powershell
separators. (Turning off srgb colorspace results in washed-out theme
colors).

--------

Side rant:
I don't agree that the recommended fix for this would be to use
`emacs-mac-port` as that basically just gives better trackpad support.
IMHO trackpad support does not outweigh the frustrations over default
meta/super key bindings, no `server-start` support, or the use of the
antiquated carbon api, but I left the recommendation as-is in this PR.