Add ESS noweb (knitr/sweave) keybindings
There are a number of functions included in ESS for dealing with knitr
and sweave "chunks" that are quite handy.

These are the bindings I use, which I humbly recommend for inclusion as
defaults.

The only potential problem that I see with their inclusion is that they
only really make sense in ess-noweb-mode (a minor mode) buffers, but
since evil-leader doesn't seem to like layering keybindings through
minor-modes, adding them to ess-mode is the best solution that I can
come up with for now.

Fix ess chunk bindings as per suggestions