Load volatile-highlights-mode
Currently it's been switched from being loaded at `:init` -> `:config`

I'm guessing this was done since recently volatile-highlights-mode is no
longer an autoloaded function. If your intention is to keep around
volatile-highlights-mode, you should add it as an autoloaded function
with `:commands` and call it during `:init`.

With the current setup, I don't think the mode is ever enabled.