Fix for rcirc load ordering and configuration
I'm not sure if this is necessary, it's possible I'm using this layer
incorrectly, but I needed to change the order in which rcirc was loaded,
and in which it was configured.

This commit fixes all the issues I was running into, but perhaps I'm not
using this as intended. Be sure to look carefully over what I changed to
make sure it makes sense for what you want this layer to do.

Remove unnecessary :commands line.