Remove company-semantic to conform with Spacemacs
Since Spacemacs set the idle delay to 0.0 for continuous completion
offering, it is too much for backend like company-semantic, and probably
other backends that require long processing time that should only be
initiated when user is actually idle. company-semantic is not compatible
with Spacemacs overall, so we should remove it otherwise it hangs
Emacs.