[Fix #992] Use gh-md for superior reading experience
With gh-md, it can use Github API to render HTML to be viewed right
inside Emacs using eww web browser. It is especially useful since it
does not required markdown-mode to be installed. One benefit is that it
automatically open the rendered buffer in other window, so for example
user is learning Clojure mode and want to read documentation of Clojure
layer, he will just simply fired up helm-spacemacs, narrow to `clojure`
and press RET. Then a nicely formatted documentation buffer appears on
the right side with all the key bindings of that layer, which greatly
enhances the explorability of Spacemacs.