remind to use fd when keybindings don't work
I found myself in <kbd>SPC : customize</kbd> without knowing how to switch out of it again when I was done, there was no menubar, and I couldn't turn it on either because the <kbd>SPC</kbd> keybindings or the Vim ones just didn't work at all.
Reading the documentation again I gave <kbd>fd</kbd> a try, and it brought the keybindings back, so I think it'd be helpful to mention that <kbd>fd</kbd> is also useful as an escape when the usual keybindings don't work.

Also wouldn't hurt if a message was displayed to use <kbd>fd</kbd> when trying to use <kbd>SPC</kbd> and it doesn't work.