osx: Set `$PATH` and `$MANPATH` from login shell
* Add spaces between `OS` and `X`
* Add `exec-path-from-shell` support

On OS X the Emacs GUI does not inherit the environment variables set by
a user's login shell.

This can lead to unexpected behaviour if the user is using brew for
example to manage additional packages.

Exec-path-from-shell resolves this issue. It executes the login shell
and by default copies `$PATH` and `$MANPATH`.