Clean state when switching to a new workspace
In i3, whenever you switch to a new workspace, it starts with a blank
frame with default wallpaper. It would be useful when we switch
workspace, the window is reset to a default buffer so it's more obvious
for first time users and is closer to i3 or Vim's tabs. Since eyebrowse
supports this feature already, we simply set it to switch to
spacemacs/home function.

Also, eyebrowse-get becomes eyebrowse--get to express that it is now
private function. So this commit fixes it appropriately.