[Fix] Disable company-tooltip-flip
Currently by default company knows when to flip the popup. However, when
the above option is set to t, it causes a strange behaviour: when the
tooltip is flipped, press down (i.e. M-n or arrow key) go up and press
up (i.e. M-p or arrow key) go down.