Add a layer for the Idris language
This removes idris-mode from the extra-langs layer and moves it to its
own layer with custom evil-leader keybindings. It also removes
documentation in extra-langs on other removed languages.

There are two major complications with idris-mode: it defines its own
evil-leader bindings which overwrite Spacemacs defaults, and it uses
the variable idris-packages, which Spacemacs uses to include the
idris-mode package, to define Idris packages to load for the current
buffer.