Fix Python and Semantic integration
When no Python executable is detected, Semantic causes an error that it cannot
retrieve include paths from the executable, causing an error that prevents
Python buffer to be switched and error is repeatedly reported whenever a Python
buffer is opened. This commit ignores the error from
semantic-python-get-system-include-path that is called when Semantic setup
itself for a Python buffer.