Add zygospore package
This package allows deleting other windows and later restore back the
old window configuration with the same key binding. For example, after
binding `C-x 1` to zygospore-toggle-delete-other-windows, we can press
`C-x 1` to temporary "zoom in" the current window by deleting other
windows and "C-x 1" again to restore it. This is especially handy when
we want to pop a shell with "C-c ;", but the shell window is small and
we want to maximize it. However, maximizing destroys current window
layout. With this package, we are free to maximize the shell window and
later restore back the old window configuration.