Give titles for documents in SPC f e h
Since Helm can candidates can also be a alist of '(DISPLAY . REAL),
where DISPLAY is the text to be displayed for a candidate when it is
listed and REAL is the actual object DISPLAY is presenting, we can use
this feature to give appropriate document titles instead of using raw
file names.

This commit also changes the way we create document source: instead
defining an alist, we properly create it with `helm-build-sync-source`,
similar to how Helm Projectile does it.

Also, we do not need to sort the file list getting from the doc/
directory just reverse the list.