Create a generic jump in buffer command
The command selects appropriate jump command based on major mode.
Currently, in Org buffer, it uses helm-org-in-buffer-headings when major
mode is Org and helm-semantic-or-imenu for everything. We do this for
consistency, so user doesn't have to make an effort to remember a
different jump key binding in Org mode, currently is SPC j.

Also remove unneeded `SPC m j` in Org just for jumping to heading, so we
can use it for something else, like a group of commands.