dotfile-template: Add version and change test
First tests for a change in the version. If this test is positive, show
a note with a link to the diff. This first result will persist through
restarts until the version is updated.

Second tests for a change between the last cached template and the
current template. Again, show a note for a positive test. This is not a
persistent note though and will disappear on restart (even if the user
does nothing).