haskell layer: do not reset haskell-process-type
`haskell-process-type` is already auto by default, there is no need to
set it again. Not setting it again makes customization for users easier,
since they can just setq `haskell-process-type` in their config
function, without having to take special steps to order the setq after
the one in the haskell configuration layer (which may, if the user is not
careful, override the user's setq).