Enabled inverted comment operators without toggle
Previously, you had to set the evilnc-invert-comment-line-by-line
to be enable inverted comments. Since whether you want to invert
comments or not varies on a case to case basis, depending on the region
you wish to comment, this is bad UI.

Made two functions per original commenting function, one for inverse
the other for regular.

Added keybindings for the new inverted operations on capital letter,
so that comment lines is SPC c l and invert comment lines is SPC c L.

I removed the comment invert toggle as it is no longer needed.

(Also removed old docs which said that SPC c r is comment region;
it is actually recompile)