Evilify page scrolling in magit
Because C-f and C-b are shadowed by h and l, it is safe to bind them to
evil page scroll. C-v and M-v, the emacs defaults for page scroll, are
broken in magit. C-d is non-functional in magit buffers and C-u is
shadowed by SPC-u, so both should also be bound to scroll.