Add binding for projectile-toggle-between-implementation-and-test
The logic for binding this to `SPC p a` is that it's analogous with
vim-projectionist's notion of "alternate" files (where alternates could
be tests, headers or what-have-you). It would be more in keeping with
Spacemacs to have it on `SPC p t`, however this is already taken by
projectile neotree. If that were to be moved to `SPC p n`, then there
would be space for it, but I have no idea whether there is stomach to
make a potentially significant change such as that.