Change shortcut of bookmarks in *spacemacs* buffer
In *spacemacs* buffer, the quick help in `[?]` shows that `m` jumps to
the menu. However, if one customizes the startup lists by setting
`dotspacemacs-startup-lists '(recents bookmarks projects)`, the
shortcut `m` of jumping to the menu would be shadowed and `m` jumps to
Bookmarks.

With the help of @travisbhartwell and @TheBB, I changed the shortcut of
jumping to the Bookmarks to `a` to avoid this shortcut conflict.