Change obsolete 'new-frame to 'make-frame
Every time the sox-layer "s-n" key is used to make a new frame, there's a warning "'new-frame' is an obsolete command (as of 22.1); use 'make-frame' instead." While 'new-frame' is aliased to 'make-frame', so there's no issue with functionality, the warning is annoying. The fix is to have "s-n" call 'make-frame'.