Attempt to delete-frame first in frame-killer
This is needed to get the frame to properly close when running a client
from the terminal. server-kill-buffer will close the frame for the
terminal only if the terminal was opened with a file argument and you
haven't changed buffers. make-frame-invisible is still needed as a fall
back if you are still working from the first emacs window since it won't
let you delete the last frame.