hybrid-mode: Cleaner implementation
Using setf is better than using the previous advice, because it was
ignoring the arguments passed to evil-insert-state and the arguments
control whether the state message displays in the minibuffer. In this
version we just switch out the function definition for
evil-insert-state, and all arguments are handled perfectly.

spacemacs-core: Don't defer hybrid-mode

hybrid-mode is now extremely lightweight, and deferring it causes
problems if someone wants to bind keys in hybrid-mode but doesn't have
it selected as their editing style on startup. The reason is the
hybrid-mode keymap will not be available and an error will be thrown.
This requires using eval-after-load in this case, and this seems overly
complicated just to avoid loading this file.