core: Add spacemacs/set-key funcs
Initial commit to implement spacemacs replacements for evil-leader
functions using built-in evil functions. The two main advantages are the
removal of the need to play around with major mode hooks and the like,
as evil-leader does, and the ability to extend the functionality of
evil-leader for spacemacs without having to hack our way around the
evil-leader implementation.

As an example of one of these advantages, I implemented
spacemacs/set-default-key which will only set the key binding if one
does not already exist. Think of this like a defvar for key bindings,
which will not clobber the user's personal choices of bindings. Using
this function in layers, will make it not matter if the user declares
the key before or after the layer has loaded.