More flexible additional packages
When the package already exists, use configuration-layer/make-package
to copy information into the existing package object.

This allows overwriting location, step and excluded, but NOT owner,
pre-layers or post-layers.

This also adds a new valid location attribute called `local-dotfile`.
This is an implementation detail, to distinguish local packages (in
layers) from packages which have their location overridden to local.
This is based on the assumption that the source code in the latter case
should not be found in a subdirectory of the owning layer, but somewhere
else entirely. For those packages, `.emacs.d/local/<pkg-name>` will be
added to the load path.