Warn Linux users away from the "xemacs" package
XEmacs is packaged in many Linux distributions as "xemacs". Some may assume that "xemacs" and "emacs" have the same relationship as "gvim" and "vim", when in reality both packages have graphical support. Since I just had to correct this misunderstanding in someone, let's put it in the README to avoid it in the future.