spacemacs-buffer: Distinguish types of updates
Add a spacemacs update button to the spacemacs buffer, and relabel the
package update button to read "Update Packages". This will hopefully
resolve some confusion about updating spacemacs vs. updating the
packages.

I also tweaked the switch-to-version function to make it a little more
verbose to use it for the spacemacs update button.