Fix debug on error toggle
Add :status option and manage :on and :off manually.  Without :status
the message is wrong (always says enabled), and with :status we need an
explicit :on and :off, because otherwise the toggle would always disable
the flag.

Not sure what’s going on here, but this one works correctly now.