holy-mode: Disable ESC key binding by default
We should not assume that users of holy mode are aware of evil and
normal state. Users were accidentally escaping into normal state with
ESC under the old implementation and didn't know how to return given
that they assumed they were only using emacs key bindings.

Fixes #3495, #3473 and #3453.

Introduce a new variable `holy-mode-allow-esc-to-normal-state` to allow
for the old behavior if users want to set this explicitly.