Move the dictionary selection code into adict hook
When `auto-dictionary-mode` is activated, it tries to guess the
dictionary no matter if a dictionary was manually selected before, what
is annoying when toggling spell-checking off/on. To avoid this, a piece
of code was added to the spell-checking toggle to select the
buffer-local dictionary if it was set.

As mentioned by lunaryorn, this part of code should also be called when
flyspell is enabled/disabled by an other mean than the spacemacs toggle.
This commit move the code to the `auto-dictionary-mode-hook` for this
purpose.