Update the FAQ for evil-leader rebinding
As `SPC L` has moved to `SPC l` and avy is now on `SPC y`, the FAQ has
to be updated because the evil-leader example doesn't make sense any
longer. The question is also moved to "How do I …" where it belongs.

The code has also been updated to be a little bit cleaner.