bind avy-goto-line to `SPC y`
After the change in 1d340dcc779f79d1767b9401b9f5a7128a7c2c33,
there was some grumbling in gitter that `SPC L` is not as convenient
as `SPC l`. Apparently avy-goto-line is a popular function.

`SPC y` is chosen because it is:

- currently unused (as far as I can tell).
- sort of mnemonic because "avy" has a "y" in it.
- doesn't require SHIFT to type
- is accessible even on non-US keyboards
  (there were suggestions such ";" and "\" that are not as accessible
   on non-US keyboards)

This change also updates the documentation regarding the change from
ace-jump-mode to avy.