(mu4e) Don't require multiple account list
This commit deactivates account selection and switching support if the
account list isn't defined. We also don't need the follow-up action of
undoing the account selection changes, so we can just skip both of the
hooks.

This is useful for users who only have 1 account - they don't need the
helm popup to select which account to use. They can just define the
variables outside of the `mu4e-account-list`.