[WIP] Add vim-like alternatives to emacs `C-c …` bindings
This commit adds `SPC .` as a prefix for `C-c bindings`, and binds a few
useful ones like `C-c C-c` and `C-c C-k` (for now). It uses a simple
macro and adding new `C-c` bindings is easy.

These bindings (i.e. `C-c C-c`, `C-c C-k`, …) are useful in some modes,
like for instance in magit to valid or cancel a commit. Vim-like
alternative bindings are missing for this, this commit try to correct
this in a general way.