idris: use popwin/motion state for special buffers
Idris development results in a lot of transient popup buffers, which
the layer currently doesn't handle particularly well. This commit uses
popwin so that these special buffers appear in a predictable location
and opens them in motion state so that they can be closed immediately by
pressing ~q~.

These changes are largely based on the Clojure layer, which has worked
very well in my experience.