declare-prefix-for-mode for all derived `clojure-mode`s.
Previously, prefixes only worked for `.clj` files, not `.cljs`, `.cljx`,
or `.cljc`. `clojure-mode.el` defines derived major modes for each of
those other filetypes.

In other parts of the clojure layer, `(dolist (m '(...` is used to apply
effects to all of the derived modes, but it was missing from the usage
of `spacemacs/declare-prefix-for-mode`.

`cider-repl-mode` was also added to this list.