Remove rainbow-delimiters-mode from mode-specific hooks
`spacemacs` now handles `rainbow-delimiters-mode` by adding it to the
`prog-mode-hook`, if wanted by the user. Some layers are adding it on
their own mode-hook, having for effect that `rainbow-delimiters-mode` is
called twice, which disable it.

This commit remove these layer-specific definitions of
`rainbow-delimiters` as it is now handled by the `spacemacs`
distribution. It also takes care of running `prog-mode-hook` in modes
that are not derived from it.

Fixes #3902