Check for smartparens package before calling its functions
smartparens does not seem to be included in the common-lisp layer.  I am using spacemacs-base with common-lisp and emacs-lisp layers, and exwm as an additional package.

Calling its functions when it isn't installed causes slime to crash before getting to a repl.  This change allows you to use slime without having smartparens installed.  If it isn't found, the hook just isn't added.