Add a check for spacemacs-{helm,ivy} exclusivity
Without this check, it's possible to activate both layer at the same
time, what could be a source of problems. This check ensure only one is
loaded. It should be removed when they will no longer be exclusive, or
when a better system to check `dotspacemacs`-variables values will be in
place.