Fix helm-source-header not updating
See issue #4605. When dotspacemacs-helm-header is changed from the
default nil, then the helm-source-header theming does not updated when
a new theme is loaded.

This was due to the helm-source-header-default-xxx variables being set
only once and not updated when a new theme is loaded. The fix adds an
advice to spacemacs/post-theme-init to update these variables.

This allows for all other code including theming to take place and then
set the variables correctly i.e

load-theme -> post-theme-init -> helm-header

This also ensures the code is correctly decoupled now that helm is one
of the optional completion layers.

When a single source helm is viewed the source face-attribute was
changed from helm-selection to default as this gives a more consistent
result for themes that have bright highlight lines and/or no helm face
values.