add faust-layer, remove faust from extra-langs
My first spacemacs layer, total elisp newb.
Does what it says in the README, with one bug:

I have auto-completion-enable-snippets-in-popup t in my .spacemacs
In faust files, only yas-keys that have a similar word in an open buffer
show up in the company menu.

Example: with no open buffers ne doesn't show a company menu,
if a file has the word nent in it, I get a menu with ne -> nentry (the snippet) and nent.

having ne$AnyTwoCharacters in the file also works, but just ne or nen doesn't.