C-j/C-k bindings in shell only with vim style
This wraps C-k/C-j binding customization in a check for the
dotspacemacs-editing-style variable, and only applies them if vim style
is selected. For hybrid or emacs style users, C-k not killing line and
C-j not sending a form of return will be confusing, and they will be
more used to C-n/p or M-n/p bindings, which exist in the shells already.