optionally enable ensime
a new variable, scala-auto-start-ensime, determines if ensime starts
when a scala file is loaded.

make scala-auto-start-ensime default to t

The current behaviour is to autostart so this will preserve it.

Adding documentation for scala-auto-start-ensime

removed a space

make scala-auto-start-ensime default to t

The current behaviour is to autostart so this will preserve it.

Adding documentation for scala-auto-start-ensime

removed a space