Remove obsolete clojure-defun-style-default-indent
clojure-defun-style-default-indent has been deprecated in favor of
clojure-indent-style. This is no longer a toggle but one of three
keywords. As it affects the way code is indented, and you should not
need to change your indentation style multiple times while editing, it
should not be a toggle. It's simple enough for the user to setq it to
desired value.

If desired, it may be added back as a completing-read selection (but I
don't think it's necessary).