Fixes powerline separators on OS X
I recently uninstalled railwaycat's emacs-mac-port (the old
recommendation for OS X) and installed @d12frosted's emacs-plus.

When I did that my powerline separators became just slightly off in
color from the rest of the spaceline. Asking on chat yielded that
ns-use-srgb-colorspace is set to true by default but should be nil.

I don't know if I put the code in the right place, but I'm open to
suggestions.