Allow user to enable rainbow-identifiers-mode without making it default
When you add the colors layer with rainbow-identifiers enabled,
currently Spacemacs always adds rainbow-identifiers-mode to the
prog-mode-hook, essentially making it the default. This adds a new variable,
colors-rainbow-identifiers-add-hook, which controls this behavior.
Defaults to t; if set to nil, this hook is not created.

Example usage has also been added to the README.org for the colors
layer.