git: remove all popup key binding for magit-dispatch-popup
The `SPC g` command prefix was to much overloaded for actions which
are already available in discoverable manner in magit. So I decided to
remove all the actions reachable from magit-dispatch-popup from `SPC g`.
magit-dispatch-popup is available under `SPC g m`. This increases the
key binding sequences but OTHO it makes everythings consistent (same
key sequence behind `SPC g s` and `SPC g m`) and we prefer consistency.

This refactoring free up a lot of precious key bindings we can use for
commands that are not available in Magit.